# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stadium SaaS API Documentation",
        "description": "API documentation for the Stadium SaaS Backend system. This API handles stadium bookings, management, and tenant operations.",
        "contact": {
            "email": "support@stadium-saas.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/v1/owner/bookings": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "List all bookings for a tenant",
                "operationId": "e04b917444e28768be5e636e4e46439f",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/owner/bookings/{id}": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "Get booking details",
                "operationId": "79a01702d67b63ccdabdc0c17f6b588b",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/owner/bookings": {
            "post": {
                "tags": [
                    "Owner"
                ],
                "summary": "Store a manual booking (Admin/Owner)",
                "operationId": "f4b160625e9c01c95d2138d07f6d52cd",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "field_id",
                                    "date",
                                    "start_time",
                                    "end_time",
                                    "customer_name",
                                    "customer_phone"
                                ],
                                "properties": {
                                    "field_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "18:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "example": "19:00"
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "0500000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created and confirmed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/owner/bookings/{id}/confirm": {
            "patch": {
                "tags": [
                    "Owner"
                ],
                "summary": "Confirm a pending booking",
                "operationId": "6571f1764dc4cb355670a547f22bdefb",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking confirmed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/owner/bookings/{id}/cancel": {
            "patch": {
                "tags": [
                    "Owner"
                ],
                "summary": "Cancel a booking",
                "operationId": "ab73080c20ac90d5e43702edc4606375",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/owner/bookings/{id}/payment": {
            "patch": {
                "tags": [
                    "Owner"
                ],
                "summary": "Record payment for a booking",
                "operationId": "24da7adcb591e0a50d9a67c424b72e6b",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_method"
                                ],
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "card",
                                            "online"
                                        ]
                                    },
                                    "payment_reference": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment recorded"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/owner/bookings/calendar": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "Get bookings for calendar view",
                "operationId": "ca5101eff658c6af2064d76473c601eb",
                "parameters": [
                    {
                        "name": "X-Tenant-Slug",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "field_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendar events"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}